move new code into evals
